+++ src/Keybinds.cpp
@@ -5,7 +5,6 @@
 using namespace geode::prelude;
 
 #ifndef GEODE_IS_IOS
-
 #include <geode.custom-keybinds/include/Keybinds.hpp>
 
 using namespace keybinds;
@@ -44,5 +43,4 @@ $execute
         log::error("Failed to get keybind manager");
     };
 };
-
 #endif