+++ changelog.md
@@ -1,4 +1,7 @@
 # v1.0.4
+- Mobile timer visibility button now changes opacity on toggle
+- Set minimum for number setting *`Scale`* to `0.25`
+- Cleaned up timer UI nodes
 
 # v1.0.3
 - Bumped Geode loader support to `v4.7.0`